requirements.txt
asgiref==3.5.0
astroid==2.11.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==36.0.2
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.4
Django==4.0.3
django-allauth==0.50.0
djangorestframework==3.13.1
docopt==0.6.2
filelock==3.6.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
oauthlib==3.2.0
Pillow==9.0.1
pipreqs==0.4.11
platformdirs==2.5.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
pylint==2.13.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.9
virtualenv==20.13.4
wrapt==1.14.0
yarg==0.1.9
